pyproject.toml
[tool.poetry]
name = "TheBCHPodBot"
version = "0.1.0"
description = ""
authors = ["JettScythe <jettscythe@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
python-telegram-bot = "^13.14"
python-dotenv = "^0.21.0"
pycoingecko = "^3.1.0"
Babel = "^2.11.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pre-commit = "^2.20.0"
commitizen = "^2.37.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
